ui/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
python-hcl2>=4.3.0
